--- apps/indexer/src/lib/viem.ts
@@ -1,1 +1,0 @@
-export {}
