--- apps/indexer/src/mods/convert/withdrawals.ts
@@ -1,1 +1,0 @@
-export {}
