+++ apps/indexer/src/modules/auction/easyauction.ts
@@ -1,5 +1,5 @@
 import { getContract } from 'viem'
-import { client } from './viem-client'
+import { client } from '../../viem-client'
 import { TestnetEasyAuction } from 'smartsale-contracts'
 
 export const easyAuction = getContract({
